Re-apply change for static_cast removal

+++ NAS2D/Resource/AnimationSet.cpp
@@ -232,7 +232,7 @@ namespace
 			reportMissingOrUnexpected(dictionary.keys(), {"sheetid", "x", "y", "width", "height", "anchorx", "anchory"}, {"delay"});
 
 			const auto sheetId = dictionary.get("sheetid");
-			const auto delay = dictionary.get<int>("delay", 0);
+			const auto delay = dictionary.get<unsigned int>("delay", 0);
 			const auto x = dictionary.get<int>("x");
 			const auto y = dictionary.get<int>("y");
 			const auto width = dictionary.get<int>("width");
@@ -274,7 +274,7 @@ namespace
 
 			const auto bounds = Rectangle<int>::Create(Point<int>{x, y}, Vector{width, height});
 			const auto anchorOffset = Vector{anchorx, anchory};
-			frameList.push_back(AnimationSet::Frame{image, bounds, anchorOffset, static_cast<unsigned int>(delay)});
+			frameList.push_back(AnimationSet::Frame{image, bounds, anchorOffset, delay});
 		}
 
 		return frameList;
